Update lecture2 tasks
I think this is really close. A file called monthly_averages was created with the dates, but the actual doubles for the calculated averages are not listed in the spreadsheet. However, they are listed  in a sheet in the workspace in a 1X30 matrix. I'm not sure why they are not showing up in the written file, because the 1X30 vector is storing new values for each element as the for-loop cycles through i.