Monthly_Stats_Function
My thinking here is that the function will begin at 1 = 2008 after setting up my variables, and then for each value of i, it will cycle through 12 j values, and for each j value, it will place the mean of that month in the first column and the standard deviation in the second column. 

However, this only gives the mean and standard deviations of each month for one year at a time.