Update lecture6_KristenWagner.m
We worked on this for a couple hours and came up with this, a triple for loop that has the goal of assigning values to a matrix. The only issue is that we can't assign a whole list of numbers to a single cell. We thought about potentially using the IDX thing but weren't sure how we'd do that without using kmeans. Help!!!